requirements.txt
# Core Dependencies
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.2

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Networking & APIs
urllib3>=1.26.0
aiohttp>=3.8.0

# STIX/TAXII Support
stix2>=3.0.0
taxii2-client>=2.3.0

# Validation & Parsing
jsonschema>=4.17.0
validators>=0.20.0
tldextract>=3.4.0

# Database (optional)
sqlalchemy>=2.0.0

# Logging & Monitoring
python-json-logger>=2.0.4
colorlog>=6.7.0

# CLI
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# SIEM Integrations
elasticsearch>=8.0.0
splunk-sdk>=1.7.0

# Caching
redis>=4.5.0
cachetools>=5.3.0

# Security
cryptography>=39.0.0

# Testing (dev dependencies)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0
responses>=0.22.0

# Code Quality (dev dependencies)
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
isort>=5.10.0
pylint>=2.15.0

# Documentation (dev dependencies)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
